Data Structure Automation
The Following Files Combine To Form a machine that automates the processing of the data structures and their operations at the dynamic time. It Gives the end user, a facility to set a size and limit operations as per the required need. 
The Mainwindow.cpp is the driver code file, which runs the actual program. However, You will have to include the .h extension file in the same destination folder.
Simply Download the Folder and Execute it.